Refactor actuator, sensor, and Firebase logic
Improves LED and buzzer control logic to reduce log spam and simplify state tracking. Refactors buzzer pattern handling for clarity and reliability. Updates battery and motion sensor data handling to use float voltage and correct axis mapping. Changes Firebase data structure to send float voltage, adds debug output for voltage transmission, and updates initialization to wait for WiFi before Firebase. Skips UART and motor board communication in debug mode for easier development.

+++ ESP32-S3-Main/src/communication/firebase_manager.h
@@ -2,14 +2,14 @@
 #define FIREBASE_MANAGER_H
 
 #include <Arduino.h>
-#include <Firebase_ESP_Client.h>
+#include <FirebaseESP32.h>
 
 struct FirebaseTxData {
     // Sensor readings - SEND ONLY
     int acceleration;           // Send acceleration
     int angular;               // Send angular velocity
     int battery;               // Send battery level
-    int voltage;               // Send voltage
+    float voltage;             // Send voltage
     int temp;                  // Send temperature
     int humidity;              // Send humidity
     int lightlevel;            // Send light amount from path LDRs
